Update README.md
My name is Frida Kahlo, and I was born on July 6, 1907, in Mexico City, Mexico. My life has been an intricate tapestry of joys and sorrows, painted with the vibrant colors of passion and suffering.

From a young age, I displayed a predisposition for art, but my path was marked by a tragic accident in 1925 when I was just an 18-year-old young woman. A terrible bus accident forever changed the course of my life. The injuries I sustained in that incident shaped the trajectory of my existence, forcing me to live with physical pain for the rest of my days.

However, suffering acted as a catalyst for my creativity. During the extended period of recovery, I began to paint self-portraits, exploring my identity and the complex emotions that illness had brought into my life. My art became a form of personal expression and a way to confront my reality.

In 1929, I married the renowned painter Diego Rivera, a tumultuous but also passionately artistic union. Our relationship was characterized by highs and lows, infidelities, and reconciliations, but it was a fundamental part of my story.

My works are often characterized by symbolic elements, surrealism, and meticulous details. I painted my life, my pains, my joys, and my struggles with a unique and unmistakable style. My art gained international recognition, and today I am considered an icon of Mexican art.

My life was prematurely interrupted on July 13, 1954, when I left this world at the age of 47. Despite my short existence, my impact on art and culture remains indelible. My story is that of a courageous woman who transformed pain into beauty, and my works continue to inspire and move people around the world.
![Frida-Khalo-padova.jpg](https://cdn-uploads.huggingface.co/production/uploads/65b11a28266937a1f9e04961/z9ZucDhx9O8BMN8bRusNz.jpeg)